// src/components/atoms/title/enums/title.component.enums.ts
enum TitleComponentFontSizeEnum {
  SMALL = 'small',
  MEDIUM = 'medium',
  LARGE = 'large',
  XLARGE = 'xlarge',
  XXLARGE = 'xxlarge',
  XXXLARGE = 'xxxlarge',
  XXXXLARGE = 'xxxxlarge',
}

enum TitleComponentFontWeightEnum {
  BLACK = 'black',
  BOLD = 'bold',
  EXTRABOLD = 'extrabold',
  EXTRALIGHT = 'extralight',
  LIGHT = 'light',
  MEDIUM = 'medium',
  NORMAL = 'normal',
  SEMIBOLD = 'semibold',
  THIN = 'thin',
}

enum TitleComponentTagEnum {
  H1 = 'h1',
  H2 = 'h2',
  H3 = 'h3',
  H4 = 'h4',
  H5 = 'h5',
  H6 = 'h6',
}

export { TitleComponentFontSizeEnum, TitleComponentFontWeightEnum, TitleComponentTagEnum };